requirements.txt
--prefer-binary
--extra-index-url https://download.pytorch.org/whl/cpu

# ----- lightweight CPU wheels (no compiler) -----
torch==2.2.2+cpu          # exists & 5× smaller than 2.8
torchvision==0.17.2+cpu   # matching pair
opencv-python-headless==4.8.1.78
Pillow==9.5.0
numpy==1.24.3
requests==2.31.0
tqdm==4.66.1
matplotlib==3.7.2
scipy==1.11.2
scikit-learn==1.3.0
fastapi==0.103.1
uvicorn[standard]==0.23.2
gradio==3.44.4            # exists (3.44.3 does not)
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.24.1
python-dotenv==1.0.0
imageio==2.31.3
#  DROP:  basicsr, gfpgan, deoldify, facexlib, realesrgan, filterpy, numba, wandb, imageio-ffmpeg → saves 5-6 min